Midterm Project
Assignment : 

As designers, we are in the crucial position to help citizens take charge of their lives and futures through finding information they are passionate about. How can we make these choices clear?Topic / Technology / Audience: We’ll generate project ideas from the brainstorming and persona exercises.You will plan, design, and build a website around your topic over the next few weeks, using the grid as a reference. 

My Project: 

This is a watered down version of other Recipe websites out there.It’s strictly for desserts, and there’s no search bar at the moment for you to find certain recipes. Users would visit this site and will be given a minimum 5 (max 10) recipes for each holiday they click on. In the future, I would be rotating the recipes, so one recipewouldn’t stay the same for too long. This will give the novice bakers a challenge and more experience with different types of pastries in their baking journey. It will also give more experienced bakers inspiring concepts for them to mimic,or for them to put a twist on.Personas: